backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.1
flask-limiter==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.3.0
numpy==1.26.4
tensorflow==2.16.1
gunicorn==22.0.0
werkzeug==3.0.3
cloudinary==1.40.0
anthropic>=0.40.0
openai>=1.0.0
google-genai>=1.0.0
pytest==8.2.0
pytest-flask==1.3.0
